Upgrade to REL4.0.x

+++ src/main/java/emu/grasscutter/server/packet/send/PacketServerGlobalValueChangeNotify.java
@@ -1,13 +1,12 @@
 package emu.grasscutter.server.packet.send;
 
 import emu.grasscutter.game.entity.GameEntity;
-import emu.grasscutter.net.packet.BasePacket;
-import emu.grasscutter.net.packet.PacketOpcodes;
+import emu.grasscutter.net.packet.*;
 import emu.grasscutter.net.proto.ServerGlobalValueChangeNotifyOuterClass.ServerGlobalValueChangeNotify;
 import emu.grasscutter.utils.Utils;
 
 public final class PacketServerGlobalValueChangeNotify extends BasePacket {
-    public PacketServerGlobalValueChangeNotify(GameEntity entity, String abilityHash, int value) {
+    public PacketServerGlobalValueChangeNotify(GameEntity entity, String abilityHash, float value) {
         super(PacketOpcodes.ServerGlobalValueChangeNotify);
 
         this.setData(
@@ -17,7 +16,7 @@ public final class PacketServerGlobalValueChangeNotify extends BasePacket {
                         .setKeyHash(Utils.abilityHash(abilityHash)));
     }
 
-    public PacketServerGlobalValueChangeNotify(int entityId, String abilityHash, int value) {
+    public PacketServerGlobalValueChangeNotify(int entityId, String abilityHash, float value) {
         super(PacketOpcodes.ServerGlobalValueChangeNotify);
 
         this.setData(
